Implement File Chunking
Due to the way that Ceph works it is important to split large files
into smaller objects, this prevents uneven utilization of storage
devices and potential performance problems.

This patch implements chunking at the 4MB mark. The last object created
in this process may be smaller than 4MB.

A header object is created which contains the following:
* The number of parts for a file
* A list of all parts, with it's size and a hash of contents
* The total size of all objects

When reading a file out of the object store we read each of it's parts
into a temporary file. Once all parts have been written that stream is
returned. This build process is done in a serial manner leading to poor
performance during reads of large files. For example, downloads will be
delayed in starting.

Signed-off-by: JuanJose "JJ" Galvez <jgalvez@redhat.com>